embedder/requirements.txt
facenet-pytorch==2.5.1
fastapi==0.63.0
gast==0.2.2
h5py==2.10.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
matplotlib==3.3.3
notebook==6.1.5
numpy==1.19.4
opencv-python==4.4.0.46
pickleshare==0.7.5
Pillow==8.0.1
pymilvus==0.2.14
pyparsing==2.4.7
pypi==2.1
pyrsistent==0.17.3
python-dateutil==2.8.1
python-multipart==0.0.5
torch==1.7.1
torchaudio==0.7.2
torchvision==0.8.2
tornado==6.1
ujson==4.0.1
urllib3==1.26.2
uvicorn==0.13.2
